Move the valid map key check into its own compiler pass.
This allows us to disable this check for tools like the message extractor, RunParser or ParamMigrator. Those tools don't have proto information, so the compiler is unable to resolve proto enums. That means that if a Soy template has a proto enum as a map key, the Soy compiler can't verify the key type is a proto enum and fails. Now, we don't check map key types for these tools.

This stashes the resolved SoyType on the TypeNode so we can nicely traverse the type tree to search for map keys, but can used the resolved type of the map key to determine if it's legal.

This also adds a build test on XMB as a simple test for the message extractor.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=194109713