Automated [] rollback of changelist 194407284.
*** Reason for rollback ***

somehow this is causing formatting issues

*** Original change description ***

Build placeholders for jssrc using the extracted message instead of the MsgNode AST

This is more similar to how it works in the other backends and is simpler imho.  Also change it to not generate the same placeholder multiple times.

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=194415541