Unwrap kind="text" keys when creating and accessing maps in Python.
This uses a similar strategy to the JS backend's maybeCoerceKeyToString functionality.

The need to do this came up in []which added a bunch of integration tests for this that didn't work in Python. They now all pass.

GITHUB_BREAKING_CHANGES=none, unless you're relying on strings and kind="text" keys to be treated differently for maps in the Python backend

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=194793929