Add support for enforcing an ordering constraint between placeholders during {msg} rendering.
For velogging, we guarantee that enterLoggableElement and exitLoggableElement are called in the correct order.  Because we will soon allow these things to exist in {msg} tags attached to placeholders, we will need to be defensive against accidental reorderings caused by bad translations.

This cl adds support and unit tests for the runtime detection.  I spent a while thinking about adding earlier detection, but it is far from trivial so for now this is a render time error.  If we see this pop up in practice we can investigate systems to flag this at build time.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=194826771