Allow {velog} in msg nodes.
This works just like {velog} in other locations except that the logonly attribute is not supported.  This is because logonly means that the content of the {velog}..{/velog} won't get rendered.  This is technically supportable with velog in a {msg} but the behavior seems bizarre (how do we know that removing part of a translation makes sense?).  So for now it is a compilation error and we recommend that people refactor their messages instead.

The implementation takes advantage of the fact that {velog} nodes must wrap a single DOM node and we already automatically turn HTML tags into placeholders.  So we can associate the logging logic with the placeholders.

* For jssrc this is trivial since we merely need to annotate the placeholders with the custom attributes (which will just happen as soon as we allow it)
* For jbcsrc this is a little trickier, we need to inject code to call enterElement and exitElement into the SoyValueProvider instances we generate for placeholders.

Additionally we need to modify all the msg code in the compiler to make it compatible with VeLogNode.
* VeLogNode needs to participate in the sameness algorithm for MsgHtmlTagNode since now two <a> tags in a message are not equivalent if one of them is in a {velog} and the other isn't
* VeLogNode needs to become a MsgBlockNode
* The algorithms that traverse MsgNode instances need to allow for VeLogNode

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=194935172