Soy JS runtime: shuffle map keys in debug mode.
This is the counterpart to [] (which is for the Java runtime).

The iteration order of Soy maps is documented as undefined.

GITHUB_BREAKING_CHANGES=The order of keys in a map (i.e., the list returned by mapKeys()) is no longer deterministic in JS in debug mode. (The order has always been documented as undefined, but until now it was deterministic.) UI produced by templates that directly iterate over map keys may become reordered. If this affects you, you need to switch to iterating over a data structure with guaranteed order (e.g. a list).

(Note that this does not affect legacy_object_maps. The order of their keys is also documented to be undefined, but in the implementation they continue to be deterministic.)

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=196532663