Soy Python runtime: maps should not have a consistent iteration order.
The iteration order of Soy `map`s is documented to be undefined, but in practice has been deterministic. To prevent customers from relying on this undocumented behavior, we changed the Java ([] and JS ([] runtimes to make the iteration order nondeterministic. This CL does the same thing for the Python runtime.

Like the Soy JS runtime, the Python runtime allows users to pass in data structures to back Soy params, without any conversion. The iteration order of map params backed by caller-supplied data is determined by that data. In the common case, if the caller passes a Python dict, the order of keys is unspecified (see https://docs.python.org/3.1/whatsnew/3.1.html).

The other source of backing data for Soy maps is map literals. Currently, the pysrc backend generates OrderedDict instances for them. This CL switches to using a simple dict.

GITHUB_BREAKING_CHANGES=The order of keys in a map (i.e., the list returned by mapKeys()) is no longer deterministic in Python. (The order has always been documented as undefined, but until now it was deterministic.) UI produced by templates that directly iterate over map keys may become reordered. If this affects you, you need to switch to iterating over a data structure with guaranteed order (e.g. a list).

(Note that this does not affect legacy_object_maps. The order of their keys is also documented to be undefined, but in the implementation they continue to be deterministic.)

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=196566384