Remove support for jspb.map_key.
This is unused and is no longer needed now that protos support maps natively.

Now any proto field annotated with jspb.map_key will be treated as a normal field. This adds a specific error message for people trying to use a jspb.map_key field as a map and directs them to use a native proto map or to reference the field as a list.

I confirmed with [](which was a [] train run of http://[] that this is no longer used.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=197005518