Add a Statement concept to the js backends
Soy generates some code that is properly a statement, e.g. a for loop or a return statement.  Right now those are modeled as CodeChunk instances, this is confusing because CodeChunk.WithValue is a subtype of CodeChunk.  This cl introduces a new subtype called Statement which will subsume these usecases.  This should make things a bit more typesafe and obvious.

Mostly this is just replacing CodeChunk with CodeChunk.Statement, however in the case of ConditionalBuilder the change was a little more intense.  We no longer support the ability to dynamicaly rewrite a ConditionalBuilder as a ternary, instead users will have to use ConditionalExpressionBuilder.  This appears to not be a concern for any users outside of the unit tests.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=197061825