Move all the Pass implementations out of PassManager and merge them with their underlying visitors.
The inline Pass implementations were added in PassManager as a way to speed up
implementing passmanager, but were never intended for long term storage. This finishes that migration.  The benefit is that I plan on adding behavior to the Pass base classes, so i want them to be the primary pass interface rather than the visitors.  This is useful for some of the backends which run passes in an adhoc manner.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=197078164