Fix a bug where slightly complex expressions for logonly statements in jbcsrc would generate bad code.
The fundamental issue is that and expression returned from SoyExpression.coerceToBoolean is not safe to gen into the method more than once.  This isn't an unusual condition for Expressions in the jbcsrc backend, the workaround is to store expression values in local variables or on the stack which is what we need to do here.

The reason we need to access the logonly expression value multiple times is in order to implement the logonly behavior for when there is no logger attached.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=198726197