Change the Pass interfaces to make only running on source files the default behavior.
Some of the more conscientious passes would internally filter to only operate on source files, but ~all passes should do this.  The easiest way is to change the interfaces.

For CompilerFilePass, we add a shouldRunOnDeps() template method and then passes that need to run on deps will override it.   Only 2 passes need to do this afaict: HtmlRewritePass and SoyConformancePass each pass contains a justification for why and a link to a bug describing how to eliminate this.

For CompilerFileSetPass none of them need to run on deps, so i changed the interface to simply accept the set of source files.

The contextual autoescaper is a special case since it doesn't use the Pass interfaces, It also mostly only needs to run on source files, however in the case of deprecated-contextual templates we do need to run on all call targets in order to calculate end contexts.  When deprecated-contextual goes away we can simplify this (and turn the autoescaper into a normal pass!).

GITHUB_BREAKING_CHANGES=The soy compiler will no longer report errors when parsing dep files (other than parser errors).  This should not affect a well behaved use of the compiler.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=198772299