(Soy Java Plugin Revamp - Step 3) - Update FunctionNode to accept with SoyFunction or SoySourceFunction. (This does not yet update the PluginResolver to return new SoySourceFunctions.)
This changes FunctionNode to store the function name + an 'Object' of the
function, instead of the typed 'SoyFunction' for the function.

We need to store the function name becaues SoySourceFunction's name is in its
annotation.  We could instead do an instanceof check in getFunctionName() and
call fn.getName() if SoyFunction or .getAnnotation(..).name() if
SoySourceFunction, but this seemed simpler & more performant. Also the other
approach won't work for the placeholder function from PluginResolver when
that is converted.

This chooses to make FunctionNode take an Object instead of creating a new
SourceFunctionNode (renaming this to LegacyFunctionNode, and introducing a
parent FunctionNode for shared functionality) because in practice, all usage
can act generically over both the legacy & new function types.  Whenever code
wants to act distinctly for the old/new types, it already had instanceof
checks for the narrower version (e.g, SoyJavaFunction), so we had to cast
anyway.

This does *NOT* actually work with the new SoySourceFunctions yet. To do that,
we need to wire the runtimes through the SoySauceImpl and change
ExpressionCompiler to use them.  That will be in a subesquent CL.

GITHUB_BREAKING_CHANGES=None

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=200592670