(Soy Java Plugin Revamp - Step 6) - Add new methods to JavaValue / JavaValueFactory.
JavaValue.soyType is necessary for functions like 'ceil' that want to do
different things based on the input type.  The switching can't be deferred
to a runtime method without requiring the param is wrapped in a SoyValue,
so it's best to expose this and let plugins try to be efficient if they can.

JavaValueFactory.listOf is necessary for functions like 'concatLists' that want
to take an arbitrary number of input args and pass a list of them to the
delegate function.  In theory this isn't strictly necessary -- we could pass
through JavaValue[] as SoyValue[]... but Lists are nicer than arrays, and
dealing with arrays would be complicated & confusing.

GITHUB_BREAKING_CHANGES=None

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=201269778