Update ways to read files.
Creating an interface that allows AbstractSoyCompiler to read multiple types of files (in this case Kythe's special database) with the read method. The read method returns a ByteSource which is then changed to a CharSource and sent as the first parameter in addWithKind function.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=206999623