Workaround the fact that our required [] plugins don't implement SoySourceFunction.
This registers them as SoyFunctions in the SoySauceBuilder so that they can be
used at runtime.

NOTE: This CL doesn't appear to be necessary anymore. With the changes from [] & reworking [] plugins to use that ([] [] is happy with the changes (at least, according to [] it is).

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=207718473