Create an ExpressionCompiler instance suitable for compiling initial @state values
Also, only allocate a static final field for the @state value if it is non-cheap.

This fixes a latent bug where a proto-init expression containing a repeated field could not be used as a state initializer and adds a minor optimization to proto-init expressions when a map or a list literal is assigned to a repeated field (common case).

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=209458722