Automated [] rollback of changelist 210190339.
*** Reason for rollback ***

if a tag is split across lines, then we may attempt to construct an invalid source location for the insertion location

*** Original change description ***

Migrate DebugSoyTemplateInfoFunction to be a builtin function

This function has always been special cased in various places around the compiler, but as part of regularizing the plugin APIs, it makes more sense to make this special function
'fully special' by making it a 'builtin'.  Builtins are soy functions
that get implemented in each backend instead of via the plugin API, this gives them access to internal compiler datastructures, which are often neccessary for them.

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=210196053