Fix a bug in how the BidiGlobalDir is calculated for SoyJavaScriptSourceFunctions
The previous approach didn't work when running the compiler in localized mode.
In that case the direction is inferred from the message bundle.  Now we share
the same logic.

To do this i refactored how we constructed the value factory so that we inject
it from the top level instead of constructing it in TranslateExprNodeVisitor

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=210643288