Rollforward cl/210649681 which was rolled back in cl/210663619 because it failed to stringify velog ids, which was rolled forward in cl/210909041 and then rolled back in cl/211011330 because the bidiMark() chars were escaped.  The rollforward fixes the escaping and adds a test
*** Reason for rollback ***

rollforward

*** Original change description ***

Automated [] rollback of changelist 210909041.

*** Reason for rollback ***

bidiMark() is escaped

*** Original change description ***

Rollforward cl/210649681 which was rolled back in cl/210663619 because it failed to stringify velog ides. The rollforward does this and asserts in the tests that velog ids are strings.

*** Original change description ***

Migrate the rest of the standard functions and a few private functions to SoyJavaScriptSourceFunction

***

***

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=211157288