Disallow converting the ve type to/from the unknown type.
In order to make sure all uses of the ve type are appropriate, we don't want to allow anyone to convert the unknown type to the ve type. We also disallow converting the ve type to unknown since there isn't a way to get the unknown type back to the ve type, so a ve type converted to the unknown type wouldn't be usable anywhere.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=220325537