Report proto init "field does not exist" errors on the location of the field name, rather than at the field value.
Previously, the field was stored as a string so we didn't have a location for it and reported the error on the closest place we did have a location, the value. http://cr/#review/198770917/depot/[]/third_party/java_src/soy/java/com/google/template/soy/exprtree/ProtoInitNode.java changed the field name to be an Identifier so now that we have a location we can report the error directly on the field name.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=223189539