Make the raw text tokens greedier.
The benefit of this is that we will require fewer tokens to parse a soy file which in turn should improve compiler performance.  Now instead of a token for every character of raw text it should be approximately one per sequence of whitespace or sequence of non-whitespace characters.

The actual benefit will depend strongly on the amount of raw text.  In the integration test suite this appears to have reduced the number of allocated bytes by about 10% and improved runtime by 3%:

https://caliper.[].com/runs/1193ced2-2fdd-42d9-ab34-30282ea3d88e,8d7c20e5-261c-470b-a65f-d0ce46668515#r:run.id,run.label

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=223216361