Adapt the ContextualAutoescaper to the CompilerFileSetPass interface
Most of this is simplifying Inferences to use a TemplateRegistry.  This also required enhancing the CompilerFileSetPass API to allow it to abort compilation completely.  This is because neither the autoescaper or the passes that run after it can deal with broken trees currently.  I took a quick look at these issues and it didn't look like it would be very hard to handle, but also relatively low value.  So the preexisting behavior is being maintained.

With this CL the autoescaper is no longer special cased in the passmanager.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=223216473