Create TemplateMetadata a substitute representation for TemplateNode in the compiler
TemplateMetadata will eventually form the basis of an object format for Soy templates to make incremental compiles possible.  This CL implements the skeleton for how this will work by changing most of the compiler to use this new API instead of TemplateNode directly.

This change adds the new class and exposes it from TemplateRegistry and then changes all TemplateRegistry users to deal with it.  In a few cases, i changed callsites to avoid TemplateRegistry altogether when it seemed like SoyFileSetNode was more appropriate.

This new API is not yet complete. Support for representing template parameters still needs to be implemented.  So for compatibility reasons i added a method getTemplateNodeForTemporaryCompatibility() to retrieve a TemplateNode from a TemplateMetadata.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=223566140