Add a representation of Template Parameters in TemplateMetadata and refactor some of the compiler to use it.
This eliminates a few more uses of getTemplateNodeForTemporaryCompatibility(), the main remaining ones exist for calculating transitive sets of parameters.  Support for this will come in a followup.

This required messing around with the 'getParamsToRuntimeCheck' feature of the Tofu runtime.  This feature makes it so parameters being passed that pass a compile time check can skip a runtime check.  Unfortunately, this feature isn't that safe in practice, but of course fixing it is out of scope for this work (and tofu is deprecated anyway).

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=223834506