Make it possible to create a CallGraph from TemplateMetadata
This is necessary in order to enable data="all" call analysis as well as to
collect information about transitive ijs.  This cl rewrites the
FindIjParamsVisitor to take advantage of the new scheme.

This cl also removes a call to getTemplateNodeForTemporaryCompatibility(),
there are only 3 remaining and all are related to FindIndirectParamsVisitor
which will be tackled next.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=223901903