Restructure the active node mechanism to use active edges.
Linking to the currently active edge of an active node when accumulating nodes doesn't work for templates that have text-only HTML. For example:
  <span>
    {if $cond1}Content{/if}
  </span>

The problem happened at the end of the IfCondNode processing. When visiting the IfCondNode, the graph cursor would correctly add a pointer to the IfCondNode to the active node list, with an TRUE active edge. It would then set the active edge to FALSE and continue processing as expected.

When finished, the accumulator then links only the FALSE edges of the IfCondNode to the synthetic phi-node.

Having a list of active edges fixes this issue. Both the node and its currently active edge are added to the active edge list. When nodes are accumulated at the end fo the IfNode, all the needed links are made.

Design doc: []
Tested:
  Added unit and functional tests as needed.

GITHUB_BREAKING_CHANGES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=224353092