Switch the VE type to store the data proto type as a string.
Before, for something like `ve(MyVe)` the data proto for `MyVe` was required to be in the proto deps even though the user never wrote a symbol for it. Now, this is no longer the case. This pushes checking the data proto type in the logging config later, when the `ve` is passed to a `ve_data` call.

This was the cause of the rollback of []
GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=225231723