Automated [] rollback of changelist 225412898.
*** Reason for rollback ***

This fails during message extraction which uses a default unknown type registry

*** Original change description ***

Add an explicit type representation in the template_metadata protos

Profiling has revealed that parsing types is the most expensive part of parsing compilation unit protos.  This is because there is a fair bit of overhead when initializing the parser for each type and because the parser actually records a lot more information than we need.  For example:
* the parser records SourceLocationInformation (which we don't use)
* the parser transforms the type strings into an AST structure and then imm...

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=225422114