Rollforward cl/225412898 which was rolled back in cl/225422114 because it broke message extraction rules since they don't configure a type registry.
*** Original change description ***
Profiling has revealed that parsing types is the most expensive part of parsing compilation unit protos.  This is because there is a fair bit of overhead when initializing the parser for each type and because the parser actually records a lot more information than we need.  For example:
* the parser records SourceLocationInformation (which we don't use)
* the parser transforms the type strings into an AST structure and then immediately rewrites it to a SoyType data structure (multiple copies)
* the parser reads from a String which entails copying every character at least twice

By recording types in a proto we can represent the same information far more succinctly and parse it faster as well.

***

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=225471633