Delay resolving types from TypeNode->SoyType until after parsing
This is part of a plan to remove 'configuration' from the parser. Which will
make it easier to write tools that only need the AST (such as an autoformatter)
and it will also make it easier to construct an AST cache if we know that no
compiler configuration needs to be in the cache keys.

As with all refactorings of compiler pass order this caused some churn in the
error message golden file tests.  This is because errors that occur during the
parse phase of the compiler prevent all further processing of that file.  To
help manage this i moved a few errors around and split one of the tests.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=225949792