Add a test that copying an AST doesn't cause any old references to be retained in the copy
As part of implementing the AST cache, I came across a difficult to diagnose bug, I eventually tracked it down to a failure to copy the TagName field of HtmlTagNode.  This meant that after a copy, if you accessed HtmlTagNode.getTagName().getNode() you would get a node from the old AST. This fixes that bug and adds a test to ensure that similar bugs don't occur in the future.

Adding the test also required fixing other bugs in AST copying.  Mostly this affected copies that occurred after running the PassManager (since various passes calculate AST sub fields and that information often wasn't copied correctly).  All bugs have been fixed except for issues the 'sameness keys'. A sameness key is an object used to tell if two placeholders in a msg are 'the same'.  These objects are typically calculated on demand (and so don't require copying logic), but in a few cases are stored in the AST.  Fixing this is a bit complex and so it is deferred.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=227550766