Add the ability to cache soy ASTs in worker mode
This is more complex than the other caches due to the unfortunate layering of AbstractSoyCompiler->SoyFileSet->SoyFileSetParser.  Because all configuration needs to be exposed on SoyFileSet.Builder we don't have a way to supply a 'pre-parsed' AST.  So we could either pass the SoyInputCache into SoyFileSetParser or tunnel it through the SoyFileSupplier API.

Because we need to read source files through the SoyCompilerFileReader interface to support kythe, we need to perform the filesystem access outside of SoyFileSetParser (which only deals with the abstracted SoyFileSupplier interface).  So i settled on using the 'tunnel an AST' approach.

It isn't very pretty but it does work.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=227554020