Fix a rare cache invalidation bug in the descriptor cache.
This is fairly unlikely to trigger, but we might as well be robust.

If a proto descriptor is provided by multiple proto_library rules (due to them having overlapping sources), then when we resolve it as a dependency we arbitrarily select one of the files to provide the dependency.  This is fine, but consider the following case.

proto_library(name='a', srcs=['a.proto', 'dep.proto'])
proto_library(name='b', srcs=['b.proto', 'dep.proto'])
proto_library(name='c', srcs=['c.proto'], deps=[':a'])

assume c.proto imports dep.proto

if all of a, b and c are provided to soy then it is possible when resolving c.proto we will look up dep.proto from 'b' instead of 'a', this will work but might cause problems if there is another invocation that only includes c and a but dep.proto has changed.  If this happens then we should invalidate 'c' but because we originally looked up 'dep.proto' from 'b' and 'b' isn't part of this compile, blaze won't tell us that b changed.  It will tell us that 'a' changed but because we think c doesn't depend on a it doesn't matter.

The solution is just for 'c' to declare that it has a dependency on both 'a' and 'b' when either could work. Then it will always get evicted if its dependency changes.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=227992536