Optimize collecting descriptors.
When doing this traversal we spend a lot of time checking whether or not we have visited a given descriptor already. The most common type registry is constructed using only FileDescriptors, in this scenario we can skip checking if we have visited a message/enum/field since we know that we will only visit each one exactly once (because we only traverse from the file down).

Also, switched to indexed for loops to save allocations.

This saves about 50% of the time of constructing a type registry.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=228214729