Cache plugin classloaders in the worker.
With all the other work in place now about 70% of the time spent in the worker is loading plugin classes.  From examining the work it is clear that it is often the same set of plugins being loaded over and over again, so the obvious solution is to add a cache.

Caching the classloaders is somewhat complex.  In order to do this effectively we need to pick the right caching granularity.  For this i decided that we should cache each af_soy_plugin rule individually.  In order to do this we need to provide more information to the compiler.  For this i have each af_soy_plugin rule produce a plugin manfest file, this is simply the list of jar files and the list of plugin class names.  These are then supplied to the worker as comma separated list.

With this the worker can now cache a classloader for each af_soy_plugin rule individually.  Also, in non-worker mode we can use this to construct a classloader much like we did before.

With this cl a hot worker can now compile in <100 ms.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=228215964