Add API to get the class name for a ServiceDescriptor.
This logic exists scattered around [], it would be good to have it in one place. http://[]

GITHUB_BREAKING_CHANGES=NONE

soy doesn't need service descriptor logic, the changes to that version of JavaQualifiedName are simply to keep it in sync with the lint cleanup in the 1P version of the class.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=228443805