For tests that require ES5, transpile Soy's JS runtime to ES5, instead of excluding files containing ES6 syntax.
This makes adding ES6 syntax to Soy's core JS runtime easier. Instead of having to go through something like []where it resulted in excluding the files contains ES6 syntax, this will automatically transpile the ES6 syntax to ES5, so it works in the cases that require ES5.

This is useful for the dynamic VE work ([]because a future CL will add more ES6 to Soy's JS runtime and this makes it so that we don't have to do anything special to get it to work with these legacy ES5 tests.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=230894838