Switch a type from string to number.
No one uses this as a string, so it'll simplify the API to switch this to a number.

This was originally a string because VE IDs were int64s and JS couldn't support that full range. Instead, this limits the value of the VE ID to be within the range JS supports.

Startblock:
  cl/232357697 is submitted

GITHUB_BREAKING_CHANGES=In JS, ElementMetadata.id is now typed as a number, instead of a string. VE IDs must now be between -9007199254740991 and 9007199254740991 (inclusive).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=232487160