Automated [] rollback of changelist 233064742.
*** Reason for rollback ***

Breaks [] []
*** Original change description ***

Switch the velog statement to use the new dynamic VE functionality.

Now, instead of hardcoding the VE ID and data into the gencode, this creates `ve` and `ve_data` objects, which it then uses to log.

This is a large change to the implementation of VE logging in Soy, but shouldn't result in any user facing changes.

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=233081428