Roll forward of []Switch the velog statement to use the new dynamic VE functionality.
Now, instead of hardcoding the VE ID and data into the gencode, this creates `ve` and `ve_data` objects, which it then uses to log.

This is a large change to the implementation of VE logging in Soy, but shouldn't result in any user facing changes.

GITHUB_BREAKING_CHANGES=none
[]_PROJECTS=[].java.com.google.image.vfe.ui,[].java.com.google.moneta.web.app.platform.ui,[].java.com.google.moneta.web.shared.loggingweb,[].java.com.google.search.[].logging.web.soyveweb,[].java.com.google.travel.frontend.destinations.uiserver

NEW: Fixed ClassCastException.

Automated [] rollback of changelist 233081428.

*** Reason for rollback ***

Roll forward, fixing the ClassCastException by correctly unboxing the VE data type. This also adds a test for the bug that caused the rollback.

*** Original change description ***

Automated [] rollback of changelist 233064742.

*** Reason for rollback ***

Breaks [] []
*** Original change description ***

Switch the velog statement to use the new dynamic VE functionality.

Now, instead of hardcoding the VE ID and data into the gencode, this creates `ve` and `ve_data` objects, which it then...

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=234153924