Add Kythe support for expression literals.
This links the map, record and ve literals to their documentation.

This moves the processing of expressions into a visitor, rather than traversing the AST multiple times to grab the different kinds of nodes. This is more efficient and easier to add a new node type (just implement the visitor method).

This adds the Identifier to the *LiteralNodes so that we can get the source location of the name of the literal.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=235916914