Allow converting SafeHtml to SanitizedHtml.
The documentation claims that Soy does it for you ([]but this is incorrect.

This method has been repeated many times:
http://[]/research/colab/notebook/onegoogle_client.py?l=147-153&rcl=236274011
http://[]/devtools/staticanalysis/pipeline/analyzers/shell/lint/templates/findings_server/findings_server.py?l=90-97&rcl=175092342
http://[]/onegoogle/appengine/oneplatform_helper.py?l=141-148&rcl=149236221
http://[]/experimental/users/gagern/ppp_overview/ppptable.py?l=131-133&rcl=184854343

This doesn't actually fix b/30781527, but means we now have the reverse conversion of [] available for use, and it de-duplicates a few different places to make it easier to migrate later.

Other tests' lint warnings addressed in []

GITHUB_BREAKING_CHANGES=None, this module is not released publicly.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=238366818