Add uses of aliases to the soyfmt tests.
This is required because it will soon be a compilation error to define an alias that's not used.

This also pipes through the original source identifier of a global reference (before it gets resolved as an alias) so we can output this same identifier in the formatter.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=238443314