Run more passes during header compilation.
It will now stop after resolving types. This is needed for default parameters, as we need the ResolveExpressionTypesPass to run in order to determine the type of the default parameter so we can set the parameter type in the header file.

This adds back the necessary inputs (plugins, logging config, globals, etc) to run the passes up to ResolveExpressionTypesPass and to resolve types on default parameter expressions.

This allows v1 expressions in Soy Header compilation mode because the Soy Header compilation is now running the V1ExpressionPass.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=238450302