SoySauce and Tofu runtime support for default parameter values.
In SoySauce this puts (non-cheap) default values in private static final fields, then references those if the parameter value is unset. This means we only have to initialize values once.

In Tofu, it's less smart because Tofu is deprecated. Any time there's a reference to a default parameter, Tofu checks if it's unset and if so evaluates and uses the default value.

This also makes default parameters work with data="all" template calls. If the caller has a default parameter that's unset, it puts the default value in the data record passed to the callee.

Implementations for the JS and Python backends will (shortly) come in later CLs.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=240578496