Fix a bug related to protos and sanitized content.
Before, for a Safe*Proto field type, the field access would just generate code to pull the wrapped value string out of the Safe*Proto. In most cases, this would get boxed later on, which would turn it into a SanitizedContent object.

But getting a Safe*Proto type from a proto and immediately setting it back in a proto wasn't getting boxed into a SanitizedContent object, so the generated code was trying to treat the wrapped value string as a SanitizedContent object.

Now, this always boxes when getting a Safe*Proto from a proto. This makes sense because a Safe*Proto field access now generates code that results in a SanitizedContent object, not a string that will hopefully get boxed later.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=241323953