Fix a bug with nested python ternary operator associativity.
Previously, the python soy compiler would compile nested `if` blocks or long
`if/ifelse` blocks into python ternary operators *without* parenthesizing any
of the subexpressions; as the ternary operator is right associative, this can
unintentionally change the semantics.

Concretely, we want a soy expression like

  {if $one}
    {if $two}
      both
    {else}
      one
    {/if}
  {else}
    {if $two}
      two
    {else}
      neither
    {/if}
  {/if}

to compile to a python expression like

  ('both' if $two else 'one') if $one else ('two' if $two else 'neither')

However, due to the lack of parens, this becomes

  'both' if $two else 'one' if $one else 'two' if $two else 'neither'

which python will interpret as

  'both' if $two else ('one' if $one else ('two' if $two else 'neither'))

which is definitely wrong.

This CL fixes the bug by parenthesizing the subexpressions in the code
generation. The cleanest solution would be to simply wrap all "print
expression" calls with something that prints parens, but due to the
particulars of how the code is written, this doesn't quite work out. Instead,
we add parens as we go, and end up tracking a number of pending right parens
to add at the end. A comment (hopefully) explains this in the code, aiding
future travelers.
Tested:
  New tests confirm several variations all work as intended; the
  original reported bug also now passes:
    []
  I also added a new integration test which fails in a clean client:
    []  and passes with this change:
    []
GITHUB_BREAKING_CHANGES=The python compiler now parenthesizes expressions in a ternary operator, eliminating a bug with short nested if/then compilation.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=241575813